emacs: overhaul {next,previous}-buffer in persps
In `perspective`, we were previously using advice to temporarily
overwrite `switch-to-prev-buffer-skip` in order to only allow switching
to buffers in the current perspective.  But this is wrong: if the user
has already configured `switch-to-prev-buffer-skip`, then we
should *compose* rather than overwrite.  This is a little complicated,
since `switch-to-prev-buffer-skip` can take on an ad-hoc set of values
that do not naturally compose.  Fortunately,
`switch-to-prev-buffer-skip` can be a function, so we can convert any
ad-hoc values into a function and then compose those functions
naturally.

- Add `ngpc/switch-to-prev-buffer-skip-function` to convert
`switch-to-prev-buffer-skip` values into functions.
- Add `ngpc/switch-to-prev-buffer-skip-all` to compose
`switch-to-prev-buffer-skip` values.
- Change the `previous-buffer`/`next-buffer` advice function in
`perspective` to use these new functions, and add advice in `persp-mode`
to use them.